refactor(worker-node): migrate WebSocket types to @disclaude/core (Issue #1041)
This PR migrates WebSocket message types to @disclaude/core as part of the
Worker Node code separation effort (Issue #1041).

## Changes

### @disclaude/core
- Add `FeishuApiAction`, `FeishuApiRequestMessage`, `FeishuApiResponseMessage` types
- Export new types from core package

### @disclaude/worker-node
- Update package version to 0.0.3
- Add ws and @types/ws dependencies (preparation for future migration)
- Update package description

### src/types/websocket-messages.ts
- Re-export all WebSocket message types from @disclaude/core for backward compatibility

### src/nodes/worker-node.ts
- Update imports to use @disclaude/core for createLogger

## Verification
- All packages build successfully
- Type check passes
- WorkerNode tests pass (5/5)

## Related
- Issue #1041 - Separate Worker Node code to @disclaude/worker-node
- Issue #1036 - WebSocket request routing (WorkerNode → PrimaryNode)
