+++ packages/worker-node/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@disclaude/worker-node",
-  "version": "0.0.2",
+  "version": "0.0.3",
   "description": "Worker Node process for disclaude - handles agent execution and scheduled tasks",
   "type": "module",
   "main": "./dist/index.js",
@@ -16,9 +16,11 @@
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "@disclaude/core": "*"
+    "@disclaude/core": "*",
+    "ws": "^8.18.0"
   },
   "devDependencies": {
+    "@types/ws": "^8.18.1",
     "typescript": "^5.9.3"
   }
 }
